+++ packages/image_bytes_cache/lib/image_bytes_cache.dart
@@ -22,6 +22,7 @@ export 'src/cache/middlewares/logger_middleware.dart';
 export 'src/cache/middlewares/skip_cache_middleware.dart';
 export 'src/http/http_bytes_client.dart';
 export 'src/http/middlewares/bearer_middleware.dart';
+export 'src/http/middlewares/conditional_middleware.dart';
 export 'src/http/middlewares/logger_middleware.dart';
 export 'src/http/middlewares/retry_middleware.dart';
 export 'src/http/middlewares/timeout_middleware.dart';