+++ packages/image_bytes_cache/test/unit_test.dart
@@ -8,6 +8,7 @@ import 'cache/skip_cache_middleware_test.dart' as skip_cache_middleware_test;
 import 'open/open_test.dart' as open_test;
 import 'resolve/bearer_middleware_test.dart' as bearer_middleware_test;
 import 'resolve/client_test.dart' as client_test;
+import 'resolve/conditional_middleware_test.dart' as conditional_middleware_test;
 import 'resolve/logger_middleware_test.dart' as logger_middleware_test;
 import 'resolve/pipeline_test.dart' as pipeline_test;
 import 'resolve/resolver_test.dart' as resolver_test;
@@ -29,6 +30,7 @@ void main() => group('Unit', () {
   client_test.main();
   pipeline_test.main();
   bearer_middleware_test.main();
+  conditional_middleware_test.main();
   logger_middleware_test.main();
   blob_store_vm_test.main();
   web_store_test.main();